+++ client/lib/components/FormChores.vue
@@ -1,11 +1,16 @@
 <script setup>
 import FormFeedback from "@vueda/components/FormFeedback.vue";
 import FormHelpText from "@vueda/components/FormHelpText.vue";
+import { useSlotNameGrouper } from "@vueda/use/useSlotNameGrouper.js";
+import { useSlotNameResolver } from "@vueda/use/useSlotNameResolver.js";
 import { THEME_OVERRIDE_PROPS, useTheme } from "@vueda/use/useTheme.js";
 import { NON_FIELD_ERRORS_KEY } from "@vueda/utils/constants.js";
+import { resolveSlotName } from "@vueda/utils/rendererSupport.js";
 import { FieldContextSymbol, FormContextSymbol } from "@vueda/utils/symbols.js";
+import { watchOnce } from "@vueuse/core";
 import omit from "lodash-es/omit.js";
-import { computed, inject, onMounted, useAttrs } from "vue";
+import pick from "lodash-es/pick.js";
+import { computed, inject, unref, useAttrs, useSlots } from "vue";
 
 const props = defineProps({
     name: {
@@ -30,13 +35,16 @@ const props = defineProps({
 });
 const formContext = inject(FormContextSymbol, null);
 const fieldContext = inject(FieldContextSymbol, null);
-onMounted(() => {
-    if (!formContext && !fieldContext && (!props.name || !props.errors || !props.warnings)) {
-        console.warn(
-            "FormChores.vue must be used within a form context, a field context, a field name must be provided with [help]/errors/warnings.",
-        );
-    }
-});
+watchOnce(
+    () => {
+        if (!formContext && !fieldContext && (!props.name || !props.errors || !props.warnings)) {
+            console.warn(
+                "FormChores.vue must be used within a form context, a field context, a field name must be provided with [help]/errors/warnings.",
+            );
+        }
+    },
+    { immediate: true },
+);
 const computedName = computed(() =>
     props.name?.length ? props.name : (fieldContext?.state?.name ?? NON_FIELD_ERRORS_KEY),
 );
@@ -66,6 +74,23 @@ const contextMessagesForName = computed(() => {
 const hasMessages = computed(
     () => Object.keys(propMessagesForName.value).length > 0 || Object.keys(contextMessagesForName.value).length > 0,
 );
+const slots = useSlots();
+const feedbackGroupSlotNames = useSlotNameGrouper(["feedback"], computedName.value, slots);
+const defaultFeedbackSlotName = computed(() =>
+    resolveSlotName(["", `feedback`], `feedback`, computedName.value, slots),
+);
+const errorFeedbackSlotName = computed(() =>
+    resolveSlotName(["error", `feedback-error`], `feedback`, computedName.value, slots),
+);
+const errorContentFeedbackSlotName = computed(() =>
+    resolveSlotName(["error-content", `feedback-error-content`], `feedback`, computedName.value, slots),
+);
+const messageFeedbackSlotName = computed(() =>
+    resolveSlotName(["message", `feedback-message`], `feedback`, computedName.value, slots),
+);
+const messageContentFeedbackSlotName = computed(() =>
+    resolveSlotName(["message-content", `feedback-message-content`], `feedback`, computedName.value, slots),
+);
 </script>
 <template>
     <div v-if="hasErrors || hasMessages || computedHelp" :class="theme('root')">
@@ -77,33 +102,51 @@ const hasMessages = computed(
                 <slot name="field-help" v-bind="slotProps" />
             </template>
         </form-help-text>
-        <form-feedback
-            v-if="hasErrors"
+        <slot
+            :name="defaultFeedbackSlotName || 'default'"
+            :has-messages="hasErrors"
             :class="theme('item')"
             :messages="props.errors"
             type="error"
             v-bind="computedAttrsSansClass"
         >
-            <template v-if="$slots[`field(${computedName})error`]" #default="slotProps">
-                <slot :name="`field(${computedName})error`" v-bind="slotProps" />
-            </template>
-            <template v-else-if="$slots[`field-error`]" #error="slotProps">
-                <slot name="field-error" v-bind="slotProps" />
-            </template>
-        </form-feedback>
-        <form-feedback
-            v-if="hasMessages"
+            <form-feedback
+                v-if="hasErrors"
+                :class="theme('item')"
+                :messages="props.errors"
+                type="error"
+                v-bind="computedAttrsSansClass"
+            >
+                <template #default="slotProps">
+                    <slot :name="errorFeedbackSlotName" v-bind="slotProps" />
+                </template>
+                <template #content="slotProps">
+                    <slot :name="errorContentFeedbackSlotName" v-bind="slotProps" />
+                </template>
+            </form-feedback>
+        </slot>
+        <slot
+            :name="defaultFeedbackSlotName || 'default'"
+            :has-messages="hasMessages"
             :class="theme('item')"
             :messages="props.warnings"
             type="message"
             v-bind="computedAttrsSansClass"
         >
-            <template v-if="$slots[`field(${computedName})message`]" #default="slotProps">
-                <slot :name="`field(${computedName})message`" v-bind="slotProps" />
-            </template>
-            <template v-else-if="$slots[`field-message`]" #message="slotProps">
-                <slot name="field-message" v-bind="slotProps" />
-            </template>
-        </form-feedback>
+            <form-feedback
+                v-if="hasMessages"
+                :class="theme('item')"
+                :messages="props.warnings"
+                type="message"
+                v-bind="computedAttrsSansClass"
+            >
+                <template #default="slotProps">
+                    <slot :name="messageFeedbackSlotName" v-bind="slotProps" />
+                </template>
+                <template #content="slotProps">
+                    <slot :name="messageContentFeedbackSlotName" v-bind="slotProps" />
+                </template>
+            </form-feedback>
+        </slot>
     </div>
 </template>